backend/requirements.txt
django>=6.0
djangorestframework>=3.15
django-cors-headers>=4.4
Pillow>=10.4
python-decouple>=3.8
gunicorn>=22.0
whitenoise>=6.7
pytest>=8.0
pytest-django>=4.8
pytest-cov>=5.0
factory-boy>=3.3
